requirements.txt
tensorflow==2.0.0
openfst-python==1.7.3
absl-py

